pyproject.toml
[project]
name = "agneticbanking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "chromadb>=1.5.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-deepseek>=1.0.1",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.1",
    "langchain-openrouter>=0.1.0",
    "langgraph>=1.0.8",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
]
